Fix the multiple recipient funding request proposal
