--- packages/ui/src/proposals/constants/regExp.ts
@@ -1,1 +1,0 @@
-export const CSV_PATTERN = /^([^,:;]+),([^,:;]+)(\n[^,:;]+,[^,:;]+)*(\n)?$/
